Dự_án/backend/requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
textblob==0.17.1
nltk==3.8.1
plotly==5.17.0
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0 
google-genai
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
httpx==0.25.2

# Database dependencies
sqlalchemy==2.0.23
mysqlclient==2.2.0
mysql-connector-python==8.2.0
alembic==1.12.1